backend/pyproject.toml
[project]
name = "deliveru-backend"
version = "1.0.0"
description = "DeliverU backend service"
authors = [
    {name = "DeliverU Team"}
]
requires-python = ">=3.12"

dependencies = [
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn>=0.30.6",
    "python-dotenv~=1.0.1",
    "python-multipart>=0.0.20",
    
    # Database
    "sqlalchemy~=2.0.35",
    "sqlmodel>=0.0.24",
    "alembic~=1.13.2",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "greenlet>=3.0.0",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "email-validator>=2.1.0.post1",
    
    # Redis & Background Jobs
    "taskiq>=0.11.17",
    "taskiq-redis>=1.0.9",
    
    # Object Storage
    "minio>=7.2.0",
    
    # Utilities
    "httpx>=0.28.1",
    "orjson~=3.10.7",
    "pyyaml~=6.0.1",
    "tenacity~=8.2.3",
]

[dependency-groups]
dev = [
    "ruff>=0.6.2",
    "mypy>=1.11.1",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.2.1",
    "types-redis>=4.6.0.20240903",
    "aiosqlite>=0.21.0",
    "pre-commit>=4.1.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W292"]

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
